feat: send RA request only in InnerClientHello when using ECH
Current implementation is inefficient as it requires copying the attestation data 4x due to the placement of the inclusion check.
It can be done much more efficiently when checking whether to write the extension at all, rather to copy it, remove it from the extension list, calculate TLSX size / write TLSX, and copy it back.
Drawback would be, that these things are not in the same place, then.
It feels a bit like a design error.